# Include position in ranking for each challenge in user page