--- a/spec/repositories/repository_challenge_spec.rb
+++ b/spec/repositories/repository_challenge_spec.rb
@@ -430,6 +430,7 @@ describe RepositoryChallenge do
         expect(challenge['best_score']).to eq(10)
         expect(challenge['best_player_score']).to eq(12)
         expect(challenge['attempts']).to eq(2)
+        expect(challenge['position']).to eq(2)
       end
     end
 
@@ -459,6 +460,7 @@ describe RepositoryChallenge do
         expect(challenge['best_score']).to eq(14)
         expect(challenge['best_player_score']).to eq(14)
         expect(challenge['attempts']).to eq(1)
+        expect(challenge['position']).to eq(1)
 
         # And challenge1 is next.
         challenge = result[1]
@@ -469,6 +471,7 @@ describe RepositoryChallenge do
         expect(challenge['best_score']).to eq(10)
         expect(challenge['best_player_score']).to eq(12)
         expect(challenge['attempts']).to eq(2)
+        expect(challenge['position']).to eq(2)
       end
     end