@Hettomei 2020-08-09
Hi.

Thanks for the PR, it s a really nice feature.

I only know mongodb's basics, I have no clue what it does in details to extract the data :) but I m concerned about performance : have you tested your request with the same amount of data as http://www.vimgolf.com/challenges/55b18bbea9c2c30d04000001 ?

Using `rake db:seed` + some params (you need to read the file), you can create such amount of data.