@filbranden 2020-08-10
Yes, performance is certainly a concern.

I think it's also not so much about a challenge with too many entries, but with a user that is in too many challenges which have too many entries...

Yes I guess I could try to create a fairly large test mongodb to take a look at that, but perhaps it would be best to test it on the real data? Any chance you could do that yourself? Or maybe could you give me access to the actual database so I can test that myself?

(Also, an option is to just push this to production and if you notice it's too slow you can rollback the commit and push it to production again...)