@filbranden 2020-08-11
If this query is indeed too expensive... Perhaps we could create a separate page `/:user/rankings` to display all the user's rankings for all challenges?

I know I would *love* to be able to see all this information together. (Curious about how many `#1`s I have...) 😁