@Hettomei 2020-08-26
After rebasing master, It works. And it s great.

My low quality benchmark:
It take 7sec to load my fake profile (with an average computer)
"entered into 51 challenges, 20 000 entry per challenge, 100 user per challenge"

I have absolutely no idea if heroku will support this, I don't know the metrics.

@igrigorik I m also curious to know if you accept the proposition of @filbranden : 

> (Also, an option is to just push this to production and if you notice it's too slow you can rollback the commit and push it to production again...)