@igrigorik 2020-08-26
9x regression would be a major hit, unfortunately I don't think that's something we can land as it'll easily turn into a DDoS vector for the site -- same problem as we had before, and I'd love not to reintroduce it.

If we want to have this functionality, we should probably explore some batch and async computation -- e.g. periodic cron job that populates a static rank table, etc. Unless, someone else has other magic to fix the 9x regression.