@filbranden 2020-09-01
Alright, so I updated this one again, to try to make the feature available while not causing an unacceptable hit on the user pages.

I did so by only displaying the ranking for each challenge and performing the expensive query if `?ranking=1` is passed to the user page URL.

(This feels somewhat like an "Easter Egg" in a way, which is not ideal... But I guess at least it allows us to push this feature to production and evaluate how it performs there, before making a decision of whether to export it more widely or even make the behavior the default.)

Let me know if you think this is acceptable for a first shot at exposing this information!

Cheers,
Filipe