@igrigorik 2020-09-13
@filbranden thanks for the nudge. Trying to grok the two different modes you're describing.. Do you have some screenshots that you can share? That might be the fastest way. :)