@filbranden 2020-09-13
So this is what the user page normally looks like:

![image](https://user-images.githubusercontent.com/106037/93031634-50c8b900-f5e1-11ea-91c0-52f3e4b27b7c.png)

It still looks the same, by default, with this PR merged.

But with this code merged in, you can add a `?ranking=1` to the URL and get this instead:

![image](https://user-images.githubusercontent.com/106037/93031650-6fc74b00-f5e1-11ea-8db6-82dcb9679587.png)

Let me highlight the differences:

![image](https://user-images.githubusercontent.com/106037/93031659-7f469400-f5e1-11ea-8e3f-77c1dde514bb.png)

The page with `?ranking=1` has a more expensive query... But that query will only be executed with the additional query parameter, which is not really exposed anywhere... So right now this is pushing a "hidden feature" of sorts, an easter egg... But this will allow us to have a good idea of how expensive this is on the production site and database, to decide whether including this information everywhere, by default, would be feasible.