@igrigorik 2020-09-14
Thanks, that's really helpful. What I was trying to wrap my head around is if/how it modifies existing pages shown by default, but based on above I think you clarified that.. it doesn't. 

I'm not opposed to shipping this as a hidden feature, but the problem is that if it does become known to folks, and they start using it, it will almost definitely affect the overall operation of the site. I guess that's a risk we could take and then pull the plug... but, that'll be painful. Tough one!