@filbranden 2020-09-15
Part of this one too is that you don't execute the expensive query *every* time you land on a user page, or when you navigate in and out of the same user page and generate it somewhat often...

Even if users find out about the "magic" query, I wouldn't expect many of them will keep checking it repeatedly (after all, it's not like you're ranking is going to go up or down that quickly...) I'd be happy if I could do it for my user once to take a snapshot :smile: 

Regarding performance considerations and possible optimization, I still think #298 (migrate to pgsql) is the way to go. I'm still interested in working on that, since I'm somewhat new to RoR and this is a great project to practice (while at the same time being one I *really* care about.)