@filbranden 2020-11-21
I tested the changes from here and I see they seem to work for users with very few (like 2) challenges:
http://www.vimgolf.com/albert07971271?ranking=1

But they seem to fail (rather quickly) for users with many challenges:

```
We're sorry, but something went wrong.
We've been notified about this issue and we'll take a look at it shortly.
```

I'm guessing it's a Heroku limit on how long it can take to build the page... So I guess that one is not as easy to solve as I expected... Oh well.